完成菜单页面显示,菜品添加,删除.

+++ Menu/src/main/java/org/zyx/repository/MenuRepository.java
@@ -20,12 +20,12 @@ public interface MenuRepository {
     @Select("select count(1) from t_menu")
     int counts();
 
-    Menu findById(int id);
+    Menu findById(long id);
 
     int save(Menu menu);
 
     int update(Menu menu);
 
-    int deleteById(int id);
+    int deleteById(long id);
 
 }
